+++ src/__tests__/boxrec.getTitleById.spec.e2e.ts
@@ -1,20 +1,17 @@
-import {CookieJar} from "request";
 import {BoxrecPageTitle} from "../boxrec-pages/title/boxrec.page.title";
 import {BoxrecPageTitlesRow} from "../boxrec-pages/titles/boxrec.page.titles.row";
 import {Boxrec} from "../boxrec.class";
 import {expectMatchDate, logIn, wait} from "./helpers";
 
-jest.unmock("request-promise");
-
 jest.setTimeout(200000);
 
 describe("method getTitleById", () => {
 
-    let loggedInCookie: CookieJar;
+    let loggedInCookie: string;
 
     beforeAll(async () => {
-        const logInResponse: { madeRequest: boolean, cookieJar: CookieJar} = await logIn();
-        loggedInCookie = logInResponse.cookieJar;
+        const logInResponse: { madeRequest: boolean, cookieString: string} = await logIn();
+        loggedInCookie = logInResponse.cookieString;
     });
 
     const WBCMiddleweightEndpoint: string = "6/Middleweight";